split message handlers
