+++ systems/networkdispatch/servermessagehandler.go
@@ -10,7 +10,6 @@ import (
 	"github.com/kkevinchou/kito/events"
 	"github.com/kkevinchou/kito/lib/network"
 	"github.com/kkevinchou/kito/managers/player"
-	"github.com/kkevinchou/kito/settings"
 )
 
 type MessageHandler func(world World, message *network.Message)
@@ -32,54 +31,6 @@ func ServerMessageHandler(world World, message *network.Message) {
 		handlePlayerInput(player, message, world)
 	} else {
 		fmt.Println("unknown message type:", message.MessageType, string(message.Body))
-	}
-}
-
-func ClientMessageHandler(world World, message *network.Message) {
-	if message.MessageType == network.MessageTypeGameStateSnapshot {
-		handleEntitySnapshot(message, world)
-	} else if message.MessageType == network.MessageTypeAckCreatePlayer {
-		handleAckCreatePlayer(message, world)
-	} else {
-		fmt.Println("unknown message type:", message.MessageType, string(message.Body))
-	}
-}
-
-func handleEntitySnapshot(message *network.Message, world World) {
-	singleton := world.GetSingleton()
-	var gameStateSnapshot network.GameStateSnapshotMessage
-	err := json.Unmarshal(message.Body, &gameStateSnapshot)
-	if err != nil {
-		panic(err)
-	}
-
-	for _, entitySnapshot := range gameStateSnapshot.Entities {
-		entity, err := world.GetEntityByID(entitySnapshot.ID)
-		if err != nil {
-			continue
-		}
-
-		cc := entity.GetComponentContainer()
-		cc.TransformComponent.Position = entitySnapshot.Position
-		cc.TransformComponent.Orientation = entitySnapshot.Orientation
-	}
-
-	for _, event := range gameStateSnapshot.Events {
-		if events.EventType(event.Type) == events.EventTypeCreateEntity {
-			var realEvent events.CreateEntityEvent
-			json.Unmarshal(event.Bytes, &realEvent)
-
-			if realEvent.EntityType == events.EntityTypeBob {
-				if realEvent.EntityID == singleton.PlayerID {
-					// TODO: skip this is event as its the creation event of the original player.
-					// In the future, the server should simply not send this message to the player
-					continue
-				}
-				bob := entities.NewBob(mgl64.Vec3{})
-				bob.ID = realEvent.EntityID
-				world.RegisterEntities([]entities.Entity{bob})
-			}
-		}
 	}
 }
 
@@ -128,30 +79,4 @@ func handleCreatePlayer(player *player.Player, message *network.Message, world W
 		EntityID:   bob.GetID(),
 	}
 	world.GetEventBroker().Broadcast(event)
-}
-
-func handleAckCreatePlayer(message *network.Message, world World) {
-	subMessage := &network.AckCreatePlayerMessage{}
-	err := json.Unmarshal(message.Body, subMessage)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	singleton := world.GetSingleton()
-	singleton.PlayerID = subMessage.ID
-	singleton.CameraID = subMessage.CameraID
-
-	bob := entities.NewBob(mgl64.Vec3{})
-	bob.ID = subMessage.ID
-
-	camera := entities.NewThirdPersonCamera(settings.CameraStartPosition, settings.CameraStartView, bob.GetID())
-	camera.ID = subMessage.CameraID
-
-	bob.GetComponentContainer().ThirdPersonControllerComponent.CameraID = camera.GetID()
-
-	world.RegisterEntities([]entities.Entity{
-		bob,
-		camera,
-	})
 }
